+++ Encoding/QSVEnc.vb
@@ -177,25 +177,43 @@ Public Class QSVEnc
             .Config = {0, 51}}
 
         Property QPI As New NumParam With {
-            .Switches = {"--cqp"},
+            .HelpSwitch = "--cqp",
             .Text = "QP I",
             .Init = 24,
             .VisibleFunc = Function() {"cqp"}.Contains(Mode.ValueText),
             .Config = {0, 51}}
 
         Property QPP As New NumParam With {
-            .Switches = {"--cqp"},
+            .HelpSwitch = "--cqp",
             .Text = "QP P",
             .Init = 26,
             .VisibleFunc = Function() {"cqp"}.Contains(Mode.ValueText),
             .Config = {0, 51}}
 
         Property QPB As New NumParam With {
-            .Switches = {"--cqp"},
+            .HelpSwitch = "--cqp",
             .Text = "QP B",
             .Init = 27,
             .VisibleFunc = Function() {"cqp"}.Contains(Mode.ValueText),
             .Config = {0, 51}}
+
+        Property QPOffsetI As New NumParam With {
+            .HelpSwitch = "--qp-offset",
+            .Text = "QP Offset I",
+            .VisibleFunc = Function() {"cqp"}.Contains(Mode.ValueText),
+            .Config = {0, Integer.MaxValue, 1}}
+
+        Property QPOffsetP As New NumParam With {
+            .HelpSwitch = "--qp-offset",
+            .Text = "QP Offset P",
+            .VisibleFunc = Function() {"cqp"}.Contains(Mode.ValueText),
+            .Config = {0, Integer.MaxValue, 1}}
+
+        Property QPOffsetB As New NumParam With {
+            .HelpSwitch = "--qp-offset",
+            .Text = "QP Offset B",
+            .VisibleFunc = Function() {"cqp"}.Contains(Mode.ValueText),
+            .Config = {0, Integer.MaxValue, 1}}
 
         Property mctf As New BoolParam With {
             .Switch = "--vpp-mctf",
@@ -269,6 +287,7 @@ Public Class QSVEnc
                         New BoolParam With {.Switch = "--b-adapt", .Text = "Adaptive B-Frame Insert"},
                         New BoolParam With {.Switch = "--i-adapt", .Text = "Adaptive I-Frame Insert"},
                         New BoolParam With {.Switch = "--adapt-ltr", .Text = "Adaptive LTR frames"},
+                        New BoolParam With {.Switch = "--extbrc", .Text = "Extended Rate Control"},
                         New BoolParam With {.Switch = "--direct-bias-adjust", .Text = "Direct Bias Adjust"},
                         New BoolParam With {.Switch = "--strict-gop", .Text = "Strict Gop"},
                         New BoolParam With {.Switch = "--open-gop", .Text = "Open Gop"})
@@ -276,7 +295,7 @@ Public Class QSVEnc
                         New NumParam With {.Switch = "--max-bitrate", .Text = "Max Bitrate", .Config = {0, Integer.MaxValue, 1}},
                         New NumParam With {.Switch = "--qp-max", .Text = "Maximum QP", .Config = {0, Integer.MaxValue, 1}},
                         New NumParam With {.Switch = "--qp-min", .Text = "Minimum QP", .Config = {0, Integer.MaxValue, 1}},
-                        New NumParam With {.Switch = "--qp-offset", .Text = "QP Offset", .Config = {0, Integer.MaxValue, 1}},
+                        QPOffsetI, QPOffsetP, QPOffsetB,
                         New NumParam With {.Switch = "--avbr-unitsize", .Text = "AVBR Unitsize", .Init = 90},
                         New BoolParam With {.Switch = "--mbbrc", .Text = "Per macro block rate control"})
                     Add("Motion Search",
@@ -308,14 +327,16 @@ Public Class QSVEnc
                         New OptionParam With {.Switch = "--videoformat", .Text = "Videoformat", .Options = {"Undefined", "NTSC", "Component", "PAL", "SECAM", "MAC"}},
                         New OptionParam With {.Switch = "--colormatrix", .Text = "Colormatrix", .Options = {"Undefined", "BT 2020 NC", "BT 2020 C", "BT 470 BG", "BT 709", "FCC", "GBR", "SMPTE 170 M", "SMPTE 240 M", "YCgCo"}},
                         New OptionParam With {.Switch = "--colorprim", .Text = "Colorprim", .Options = {"Undefined", "BT 709", "SMPTE 170 M", "BT 470 M", "BT 470 BG", "SMPTE 240 M", "Film", "BT 2020"}},
-                        New OptionParam With {.Switch = "--colorrange", .Text = "Colorrange", .Options = {"Undefined", "Limited", "TV", "Full", "PC", "Auto"}},
                         New OptionParam With {.Switch = "--transfer", .Text = "Transfer", .Options = {"Undefined", "BT 709", "SMPTE 170 M", "BT 470 M", "BT 470 BG", "SMPTE 240 M", "Linear", "Log 100", "Log 316", "IEC 61966-2-4", "BT 1361 E", "IEC 61966-2-1", "BT 2020-10", "BT 2020-12", "SMPTE 2084", "SMPTE 428", "ARIB-SRD-B67"}},
+                        New OptionParam With {.Switch = "--atc-sei", .Text = "ATC SEI", .Init = 1, .Options = {"Undef", "Unknown", "Auto", "Auto_Res", "BT 709", "SMPTE 170 M", "BT 470 M", "BT 470 BG", "SMPTE 240 M", "Linear", "Log 100", "Log 316", "IEC 61966-2-4", "BT 1361 E", "IEC 61966-2-1", "BT 2020-10", "BT 2020-12", "SMPTE 2084", "SMPTE 428", "ARIB-STD-B67"}},
+                        New OptionParam With {.Switch = "--colorrange", .Text = "Colorrange", .Options = {"Undefined", "Limited", "TV", "Full", "PC", "Auto"}},
                         MaxCLL, MaxFALL, Chromaloc,
                         New BoolParam With {.Switch = "--pic-struct", .Text = "Set the picture structure and emits it in the picture timing SEI message"},
                         New BoolParam With {.Switch = "--fullrange", .Text = "Fullrange"},
                         New BoolParam With {.Switch = "--aud", .Text = "AUD"})
                     Add("Input/Output",
                         New StringParam With {.Switch = "--input-option", .Text = "Input Option"},
+                        New OptionParam With {.Switch = "--input-csp", .Text = "Input CSP", .Init = 2, .Options = {"Invalid", "NV12", "YV12", "YUV420P", "YUV422P", "YUV444P", "YUV420P9LE", "YUV420P10LE", "YUV420P12LE", "YUV420P14LE", "YUV420P16LE", "P010", "YUV422P9LE", "YUV422P10LE", "YUV422P12LE", "YUV422P14LE", "YUV422P16LE", "YUV444P9LE", "YUV444P10LE", "YUV444P12LE", "YUV444P14LE", "YUV444P16LE"}},
                         New OptionParam With {.Switch = "--interlace", .Text = "Interlace", .Options = {"Undefined", "TFF", "BFF"}})
                     Add("Other",
                         New StringParam With {.Text = "Custom", .Quotes = QuotesMode.Never, .AlwaysOn = True},
@@ -433,7 +454,14 @@ Public Class QSVEnc
                 Case "qvbr-q"
                     ret += " --qvbr-q " & CInt(Quality.Value) & " --qvbr " & p.VideoBitrate
                 Case "cqp"
-                    ret += " --" + Mode.ValueText + " " & CInt(QPI.Value) & ":" & CInt(QPP.Value) & ":" & CInt(QPB.Value)
+                    ret += " --cqp " & CInt(QPI.Value) & ":" & CInt(QPP.Value) & ":" & CInt(QPB.Value)
+
+                    If QPOffsetI.Value <> QPOffsetI.DefaultValue OrElse
+                        QPOffsetP.Value <> QPOffsetP.DefaultValue OrElse
+                        QPOffsetB.Value <> QPOffsetB.DefaultValue Then
+
+                        ret += " --qp-offset " & CInt(QPOffsetI.Value) & ":" & CInt(QPOffsetP.Value) & ":" & CInt(QPOffsetB.Value)
+                    End If
                 Case Else
                     ret += " --" + Mode.ValueText + " " & p.VideoBitrate
             End Select